Remove file:// prefix from image path for GNTP
ruby_gtnp supports embedding image data directly in the notification
allowing remote notifications to display an image.

In practice, this change has been tested with Growl for Windows
receiving the notification and it behaves as expected, i.e. the image
appears in the notification.

This also likely resolves issue #231, as that involves Growl (for OSX)
refusing to download the images.