better emacs notifier
Wrote some elisp that does a few things:

In `Notiffany::Notifier::Emacs::Client#initialize`, runs some elisp to
define a function ( but only if it isn't already defined ). This
function takes two arguments: `color`, and `bgcolor`. The function
that's defined does a few extra things:

1. it records the current foreground and background colours of the
mode-line face
2. it sets the mode-line face to what was passed in
3. it sets a timeout to reset the mode-line face back to what it was originally

In the actual `notify` call, it runs the function that was defined in
`initialize` in order to set the mode-line colors.