Fix notifysend check when "which" is unavailable
Recently ran into this problem when building my application into a
docker container with a minimal environment.  The `which` utility
wasn't available to my application, so the return value of Sheller's
`stdout` call was nil, which caused the `empty?` check to blow up.

This adds in a nil check as well.